requirements.txt
Django==1.11.5
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-crispy-forms==1.6.1
easy-thumbnails==2.4.2
olefile==0.44
Pillow==4.2.1
pytz==2017.2
